// Facturas_Angular/Facturas/src/app/API/models/cliente.ts
/* tslint:disable */
/* eslint-disable */
export interface Cliente {
  codigoPostal?: null | string;
  domicilio?: null | string;
  fechaAlta?: Date;
  id?: number | null;
  nif?: null | string;
  nombre?: null | string;
  pais?: null | string;
  poblacion?: null | string;
  provincia?: null | string;
}